swell-coca: fix Metal issue where an occluded window ends up stale on eventual show on macOS 12/M1 -- from b37893f9

+++ WDL/swell/swell-dlg.mm
@@ -4361,11 +4361,24 @@ void swell_updateAllMetalDirty() // run from a timer, or called by UpdateWindow(
   if (r) return;
   r=true;
 
+  NSWindow *lw = NULL;
+  bool lw_occluded = false;
+
   int x = s_mtl_dirty_list.GetSize();
   while (--x>=0)
   {
     SWELL_hwndChild *slf = s_mtl_dirty_list.Get(x);
     if (!slf) break; // deleted out from under us?!
+
+    NSWindow *w = [slf window];
+    if (lw != w)
+    {
+      // this might only be needed on macOS12+ and/or M1 (could not duplicate on high sierra)
+      lw = w;
+      NSUInteger (*send_msg)(id, SEL) = (NSUInteger (*)(id, SEL))objc_msgSend;
+      lw_occluded = w && !(send_msg(w, @selector(occlusionState))&2);
+    }
+    if (lw_occluded) continue;
 
     const RECT tr = slf->m_metal_in_needref_rect;
     s_mtl_dirty_list.Delete(x);
